Linux fixes

+++ vm/vm.cpp
@@ -6,6 +6,8 @@
 #include "state.hpp"
 #include "disassembler.hpp"
 
+#include <stdio.h>
+
 #define TRACE
 
 namespace marius {
@@ -235,7 +237,7 @@ namespace marius {
       }
     }
 
-    for(int i = 0; i < args.size(); i++) {
+    for(size_t i = 0; i < args.size(); i++) {
       fp[i] = temp[i];
     }
   }
